refactor: Move parameterizer to transformer package
