+++ cmd/move2kube/move2kube.go
@@ -66,7 +66,6 @@ For more documentation and support, visit https://move2kube.konveyor.io/
 	rootCmd.AddCommand(getCollectCommand())
 	rootCmd.AddCommand(getPlanCommand())
 	rootCmd.AddCommand(getTransformCommand())
-	rootCmd.AddCommand(getParameterizeCommand())
 
 	assetsFilePermissions := map[string]int{}
 	err := yaml.Unmarshal([]byte(assets.AssetFilePermissions), &assetsFilePermissions)
