--- lib/testdata/parameterizers/pack.yaml
@@ -1,9 +1,0 @@
-apiVersion: move2kube.konveyor.io/v1alpha1
-kind: Packaging
-spec:
-  helm: "helm-chart"
-  kustomize: "kustomize"
-  openshiftTemplates: "openshift-template"
-  parameterizerRefs:
-    - p1
-    - p2
