+++ transformer/kubernetes/parameterizer/testdata/k8s-resources/dep-v1beta1.yaml
