+++ transformer/kubernetes/parameterizer/testdata/k8s-resources/namespace-v1.yaml
