+++ transformer/kubernetes/parameterizer/testdata/parameterizers/p2.yaml
