+++ transformer/kubernetes/parameterizer/testdata/want/helm-chart/myproject/templates/dep-v1.yaml
