+++ transformer/kubernetes/parameterizer/testdata/want/helm-chart/myproject/templates/dep-v1beta1.yaml
