+++ transformer/kubernetes/parameterizer/testdata/want/helm-chart/myproject/templates/namespace-v1.yaml
