+++ transformer/kubernetes/parameterizer/testdata/want/kustomize/base/dep-v1.yaml
