+++ transformer/kubernetes/parameterizer/testdata/want/kustomize/base/dep-v1beta1.yaml
