+++ transformer/kubernetes/parameterizer/testdata/want/kustomize/base/kustomization.yaml
