+++ transformer/kubernetes/parameterizer/testdata/want/kustomize/base/namespace-v1.yaml
