+++ transformer/kubernetes/parameterizer/testdata/want/kustomize/overlays/dev/apps-v1-deployment-nginx.yaml
