+++ transformer/kubernetes/parameterizer/testdata/want/kustomize/overlays/dev/extensions-v1beta1-deployment-javaspringapp.yaml
