+++ transformer/kubernetes/parameterizer/testdata/want/kustomize/overlays/prod/apps-v1-deployment-nginx.yaml
