+++ transformer/kubernetes/parameterizer/testdata/want/kustomize/overlays/prod/extensions-v1beta1-deployment-javaspringapp.yaml
