+++ transformer/kubernetes/parameterizer/testdata/want/kustomize/overlays/prod/kustomization.yaml
