+++ transformer/kubernetes/parameterizer/testdata/want/kustomize/overlays/staging/apps-v1-deployment-nginx.yaml
