+++ transformer/kubernetes/parameterizer/testdata/want/kustomize/overlays/staging/extensions-v1beta1-deployment-javaspringapp.yaml
