+++ transformer/kubernetes/parameterizer/testdata/want/kustomize/overlays/staging/kustomization.yaml
