+++ transformer/kubernetes/parameterizer/testdata/want/m2kconfig.yaml
