+++ transformer/kubernetes/parameterizer/testdata/want/m2kqacache.yaml
