+++ transformer/kubernetes/parameterizer/testdata/want/openshift-template/parameters-dev.yaml
