+++ transformer/kubernetes/parameterizer/testdata/want/openshift-template/parameters-prod.yaml
