+++ transformer/kubernetes/parameterizer/testdata/want/openshift-template/parameters-staging.yaml
