+++ transformer/kubernetes/parameterizer/testdata/want/openshift-template/template.yaml
